Move Copilot menu card test
